Move `getGovActionLifetime` to `Query.hs`
